refactored user_agent into a new class
