+++ percy/user_agent.py
@@ -1,0 +1,46 @@
+# from IPython import embed
+import re
+import percy
+
+class UserAgent(object):
+
+    def __init__(self, client):
+      self.client = client
+
+    def __str__(self):
+        client = ' '.join(filter(None, [
+          "Percy/%s" % self._api_version(),
+          "python-percy-client/%s" % self._client_version(),
+        ]))
+
+        environment = '; '.join(filter(None, [
+          self._environment_info(),
+          "python/%s" % self._python_version(),
+          self.client.environment.current_ci,
+        ]))
+
+        return "%s (%s)" % (client, environment)
+
+    def _client_version(self):
+        return percy.__version__
+
+    def _python_version(self):
+        try:
+            from platform import python_version
+            return python_version()
+        except ImportError:
+            return 'unknown'
+
+    def _django_version(self):
+        try:
+            import django
+            return "django/%s" % django.get_version()
+        except ImportError:
+            return None
+
+    def _api_version(self):
+        return re.search('\w+$', self.client.config.api_url).group(0)
+
+    def _environment_info(self):
+        # we only detect django right now others could be added
+        return self._django_version()
